Hopefully working npm type defs
