+++ .gitignore
@@ -3,3 +3,4 @@ tests/hydrus/
 tests/hydrus-v???.tar.zst
 comparison.tmp.json
 package-lock.json
+defs/
