+++ CHANGELOG.UNCOMMITTED.MD
@@ -2,28 +2,11 @@
 
 ## <!-- version -->
 
-<!--
-    DO NOT change the insert_point and version lines above.
-    This note, "_Optional notice_" (if unchanged), and any unused groups should be removed before release.
-    Reference https://common-changelog.org/ for formatting.
-    Make sure to attribute all authors. Ideally by linking to their GitHub profile with their name as the text.
-    [shadownetdev1](https://github.com/shadownetdev1) for example.
--->
+## Changed
 
-_Optional notice_
-
-### Changed
-
-- N/A
-
-### Added
-
-- N/A
-
-### Removed
-
-- N/A
+- All node package versions are now pinned
 
 ### Fixed
 
-- N/A
+- Type defs should now work when used as a node module
+- Minor bugs in the test file
